ConnectionPool changes
Broke the pool's behaviour up into smaller chunks for reasoning
Added more try/catch stuff
OperationCanceledException will be used to determine, funnily enough, whether the current operation is canceled (for cleanup)

If the pool is full and all the connections are in use, a reservation request will be added to a queue. If a connection re-enters the pool (by being released or recreated), then the reservation request will be fulfilled. This is a little nicer than the previous behaviour of looping and then waiting 5ms before trying again to get a connection.