Replace both ConcurrentQueue<T> with single BlockingCollection<T>
Bit nicer, avoids the situation where the result is a request hanging on the request queue, and an idle connection hanging on the idle queue (due to concurrency issues).
Supports both blocking and non-blocking add/take functionality.

Changed LastActive property to be derived from when (approx.) the underlying socket last sent or received data