Final changes for the AseCommandBuilder.
Implemented and tested CommandBehavior.SingleResult and CommandBehavior.SingleRow. The implementation is naive and doesn't stop the server returning more data than necessary - it just stops the AseDataReader from processing the extraneous records.

Implemented the CommandBehavior.SchemaOnly and CommandBehavior.KeyInfo options to efficiently get the schema info and key info for use with the AseCommandBuilder.

Dropped the AseDbType.Float as this was confusing things and The SAP driver doesn't support it.

Tweaked ValueReader to return DBNull.Value when the stream indicates that data is non-existent. null should indicate an uninitialised reference.

Added support to the AseCommandBuilder to DeriveParameters from a procedure name. Similar to the SqlCommandBuilder the SAP client supports this.