Refactor in preparation for introducing an SslStream.
Changed TokenStream to decorate NetworkStream instead of extend it. This separation will allow us to inject an SslStream in between them down the track.

Need to figure out some way of sending the packets as bytes - but the packet headers complicate things as they include the packet Type and Status.