More refactoring to support encryption.
Backed out introduction of IBodyPacket.

Moved packet logic out of RegularSocket and into TokenStream. By using a NetworkStream as the source of the data, and a TokenStream to process it, we will be able to inject an SslStream in between to support encryption.

Renamed ITokenParser to ITokenReader which better described what it did in the nomenclature of streams.

Still to do - rework RegularSocket so that the NetworkStream, and later the SslStream is long-lived. Currently the Socket lives a long time, but we don't want to re-negotiate an SslStream each time we read or write.