Worked the SslStream into the InternalConnectionFactory.
Added support for validating the server certificate by either looking at the file specified in the TrustedFile connection string property, or the root X509Store. This is a point of deviation from the SAP AseClient.

Need to profile the SAP AseClient to see how it behaves in some error scenarios if we are going to be compatible.